requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pillow==12.1.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
